+++ src/compiler.js
@@ -4,9 +4,10 @@ const { execRegexOnAll } = require('./render-utils.js');
 /**
  * @param {String} abellTemplate
  * @param {Object} sandbox
+ * @param {Object} options
  * @return {String}
  */
-function compile(abellTemplate, sandbox) {
+function compile(abellTemplate, sandbox, options) {
   // Finds all the JS expressions to be executed.
   const { matches, input } = execRegexOnAll(/\\?{{(.+?)}}/gs, abellTemplate);
   let renderedHTML = '';
@@ -15,29 +16,21 @@ function compile(abellTemplate, sandbox) {
   for (const match of matches) {
     // Loops Through JavaScript blocks inside '{{' and '}}'
     let value = '';
+    const lineOffset = (input.slice(0, match.index).match(/\n/g) || []).length;
     if (match[0].startsWith('\\{{')) {
       // Its a comment! e.g. \{{ print this as it is }}
       // Ignore the match that starts with slash '\' and return the same value without slash
       value = match[0].slice(1);
-    } else if (match[0].match(/\{{.\s*}}/)) {
+    } else if (match[0].match(/\{{\s*}}/)) {
       // removes empty brackets.
       // e.g. <div>{{ }}</div> renders <div></div>
       value = '';
-    } else if (match[1].match(/} ?=/g) !== null) {
-      // Condition to check if the block has destructuring
-
-      // destructured elements need to be executed line by line.
-      const lines = match[1]
-        .trim()
-        .split(/[\n;]/)
-        .filter((line) => line.trim() !== '');
-
-      for (const line of lines) {
-        ({ sandbox } = execute(line, sandbox));
-      }
     } else {
       // Executes the block directly
-      const executionInfo = execute(match[1], sandbox);
+      const executionInfo = execute(match[1], sandbox, {
+        ...options,
+        lineOffset
+      });
       if (executionInfo.type === 'assignment') {
         sandbox = executionInfo.sandbox;
       } else if (executionInfo.type === 'value') {